+++ CHANGELOG.md
@@ -3,18 +3,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Changelog](http://keepachangelog.com/).
 
-
 ## Unreleased
+
 ---
 
 ### New
+
+- [#1123](https://gitlab.com/meltano/meltano/issues/1123) Add first-class "Submit Issue" CTA to help expedite resolution when a running job fails. Also updated the "Log" CTA in the Pipelines UI to reflect a failed state.
 
 ### Changes
 
 ### Fixes
 
 ### Breaks
-
 
 ## 1.2.0 - (2019-10-21)
 
@@ -37,7 +38,6 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - [#1050](https://gitlab.com/meltano/meltano/issues/1050) Fix a bug where the Job log would be created before the `transform` are run.
 - [#1122](https://gitlab.com/meltano/meltano/issues/1122) `meltano elt` will now properly run when using `target-snowflake`.
 - [#1159](https://gitlab.com/meltano/meltano/issues/1159) Minor UI fixes (proper `MainNav` Model icon active color during Analyze route match & "Run" auto query related cleanup) and `...NameFromRoute` refactor renaming cleanup
-
 
 ## 1.1.0 - (2019-10-16)